+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- DateFilter component
+- Link to edit register on report screen
+
 ## [0.4.0] - 2020-10-28
 
 ### Added
@@ -33,4 +37,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Registers history
 - Delete register
 - Update register
-